# documentacion/stock_control-openapi.yaml
openapi: "3.0.3"
info:
  title: "stock_control API"
  description: "stock_control API"
  version: "1.0.0"
servers:
  - url: "https://stock_control"
paths:
...
  /categoria/{id}:
    get:
      summary: "GET categoria/{id}"
      operationId: "categoryDetail"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
    put:
      summary: "PUT categoria/{id}"
      operationId: "categoryUpdate"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
    delete:
      summary: "DELETE categoria/{id}"
      operationId: "categoryDelete"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
...
  /usuario/{id}:
    get:
      summary: "GET usuario/{id}"
      operationId: "userDetail"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
    put:
      summary: "PUT usuario/{id}"
      operationId: "userUpdate"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
    delete:
      summary: "DELETE usuario/{id}"
      operationId: "userDelete"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
  /usuarios:
    get:
      summary: "GET usuarios"
      operationId: "userList"
      parameters:
        - name: "page"
          in: "query"
          required: true
          schema:
            type: "number"
            format: "int32"
      responses:
        "200":
          description: "OK"